feat(chips): define constant - cssClasses
