+++ packages/chips/chip/constant.ts
@@ -1,6 +1,6 @@
 // The MIT License
 //
-// Copyright (c) 2018 Google, Inc.
+// Copyright (c) 2019 Google, Inc.
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -19,36 +19,20 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-import React from 'react';
 
-export default class ChipCheckmark extends React.Component<{}, {}> {
-  width: number = 0;
+import {cssClasses} from '@material/chips/chip/constants';
 
-  init = (element: HTMLDivElement) => {
-    if (!element) {
-      return;
-    }
-    // The checkmark's width may initially be set to 0, so use the checkmark's height as a proxy since the
-    // checkmark should always be square.
-    this.width = element.getBoundingClientRect().height;
-  };
+const CSS_CLASSES = {
+  ...cssClasses,
+  CHECKMARK_PATH: 'mdc-chip__checkmark-path',
+  CHECKMARK_SVG: 'mdc-chip__checkmark-svg',
+  EXIT: cssClasses.CHIP_EXIT,
+  ICON: 'mdc-chip__icon',
+  ICON_LEADING: cssClasses.LEADING_ICON,
+  ICON_LEADING_HIDDEN: cssClasses.HIDDEN_LEADING_ICON,
+  ICON_TRAILING: cssClasses.TRAILING_ICON,
+  ROOT: 'mdc-chip',
+  TEXT: 'mdc-chip__text',
+};
 
-  render() {
-    return (
-      <div className='mdc-chip__checkmark' ref={this.init}>
-        <svg
-          className='mdc-chip__checkmark-svg'
-          viewBox='-2 -3 30 30'
-          focusable='false'
-        >
-          <path
-            className='mdc-chip__checkmark-path'
-            fill='none'
-            stroke='black'
-            d='M1.73,12.91 8.1,19.28 22.79,4.59'
-          />
-        </svg>
-      </div>
-    );
-  }
-}
+export {CSS_CLASSES};
