+++ test/unit/chips/ChipSet.test.tsx
@@ -4,7 +4,7 @@ import td from 'testdouble';
 import {shallow, mount} from 'enzyme';
 import ChipSet from '../../../packages/chips/chip-set/ChipSet';
 import {Chip, ChipProps} from '../../../packages/chips/index'; // eslint-disable-line @typescript-eslint/no-unused-vars
-import ChipCheckmark from '../../../packages/chips/ChipCheckmark';
+import ChipCheckmark from '../../../packages/chips/chip/ChipCheckmark';
 import {coerceForTesting} from '../helpers/types';
 
 suite('ChipSet');
